refactor: InputUtil 추가 분리 작업 진행
- 추가 숫자 문자열 반환 로직 InputUtil 이동
